refactor textract
