+++ zoning/ocr/textract.py
@@ -1,0 +1,154 @@
+import json
+import os
+import time
+from typing import Generator, Tuple
+
+import boto3
+import tqdm
+from omegaconf import DictConfig
+from tqdm.contrib.concurrent import process_map, thread_map
+
+from class_types import Entities, Entity, ExtractionTargetCollection, ExtractionTarget
+from .base_extractor import Extractor
+
+class TextractExtractor(Extractor):
+    def __init__(self, extractor_config: DictConfig):
+        super().__init__(extractor_config)
+        if self.config.extract.run_ocr:
+            self.extractor = boto3.client("textract")
+
+    def start_job(self, s3_bucket_name: str) -> str:
+        """Runs Textract's StartDocumentAnalysis action and specifies an s3
+        bucket to dump output."""
+        response = self.extractor.start_document_analysis(
+            DocumentLocation={
+                "S3Object": {
+                    "Bucket": self.config.extract.input_document_s3_bucket,
+                    "Name": s3_bucket_name,
+                }
+            },
+            FeatureTypes=self.config.extract.feature_types,
+        )
+
+        return response["JobId"]
+
+    def get_job_status(self, job_id: str) -> Generator[Tuple[str, str], None, None]:
+        """' Checks whether document analysis still in progress."""
+        status: str = "IN_PROGRESS"
+        while status == "IN_PROGRESS":
+            time.sleep(5)
+            response = self.extractor.get_document_analysis(JobId=job_id)
+            status = response["JobStatus"]
+            yield status, response.get("StatusMessage", None)
+
+    def get_job_results(self, job_id: str) -> Generator[dict, None, None]:
+        """If document analysis complete, runs Textract's GetDocumentAnalysis
+        action and pulls JSON results to be stored in s3 bucket designated
+        above."""
+        response = self.extractor.get_document_analysis(JobId=job_id)
+        nextToken = response.get("NextToken", None)
+        yield response
+
+        while nextToken is not None:
+            response = self.extractor.get_document_analysis(
+                JobId=job_id, NextToken=nextToken
+            )
+            nextToken = response.get("NextToken", None)
+            yield response
+            
+    def collect_relations(self, w) -> list[str]:
+        rels = w["Relationships"] if "Relationships" in w else []
+        ids = []
+        for r in rels if rels else []:
+            for id in r["Ids"]:
+                ids.append(id)
+        return ids
+
+    def _extract(self, target: ExtractionTarget) -> None:
+        if self.config.extract.pdf_name_prefix_in_s3_bucket:
+            s3_bucket_name = self.config.extract.pdf_name_prefix_in_s3_bucket + target.pdf_file
+        else:
+            s3_bucket_name = target.pdf_file
+        job_id = self.start_job(s3_bucket_name)
+        for s in self.get_job_status(job_id):
+            status, status_message = s
+            if status == "FAILED":
+                print(
+                    f"Job {job_id} on file {target.pdf_file} FAILED. Reason: {status_message}"
+                )
+            elif status == "SUCCEEDED":
+                result = list(self.get_job_results(job_id))
+                with open(target.ocr_result_file, "w", encoding="utf-8") as f:
+                    json.dump(result, f)
+                print(f"Job {job_id} on file {target.pdf_file} SUCCEEDED. Write to {target.ocr_result_file}")
+
+    def extract(self, extract_target: ExtractionTargetCollection) -> None:
+        if self.config.extract.run_ocr:
+            thread_map(self._extract, extract_target)
+        
+        # OCR result from textract is a list of json objects containing unneeded information, we need to extract the text from it
+        self.process_ocr_results(extract_target)
+    
+    def _process_ocr_result(self, target: ExtractionTarget) -> None:
+        """
+        Inputs:
+            town (string): name of town whose text data to import
+        Returns: TODO
+        """
+        with open(target.ocr_result_file, "r") as f:
+            try:
+                data = json.load(f)
+            except json.JSONDecodeError:
+                print(f"Error decoding {target.ocr_result_file}")
+                return None
+
+        extract_blocks = [b for d in data for b in d["Blocks"]]
+
+        entities = Entities([], set(), {})
+        rows = []
+        for w in tqdm.tqdm(extract_blocks):
+            if w["BlockType"] in ["LINE", "WORD", "CELL", "MERGED_CELL"]:
+                e = Entity(
+                    w["Id"],
+                    w.get("Text", ""),
+                    w["BlockType"],
+                    self.collect_relations(w),
+                    (
+                        (w["RowIndex"], w["ColumnIndex"])
+                        if "RowIndex" in w and "ColumnIndex" in w
+                        else (-1, -1)
+                    ),
+                )
+                entities.add(e)
+            elif w["BlockType"] == "PAGE":
+                if len(entities.ents) > 0:
+                    rows.append(
+                        {
+                            "Town": f"{target.name}",
+                            "Page": w["Page"] - 1,
+                            "Text": str(entities),
+                        }
+                    )
+                entities = Entities([], set(), {})
+            elif w["BlockType"] == "TABLE":
+                pass
+            else:
+                continue
+
+        if len(entities.ents) > 0:
+            rows.append(
+                {
+                    "Town": f"{target.name}",
+                    "Page": w["Page"],
+                    "Text": str(entities),
+                }
+            )
+
+        with open(target.dataset_file, "w") as f:
+            json.dump(rows, f)
+       
+    def process_ocr_results(self, extract_target: ExtractionTargetCollection) -> None:
+        assert len(os.listdir(extract_target.ocr_result_dir)) > 0, "No OCR results found"
+        process_map(self._process_ocr_result, extract_target.targets)
+
+
